+++ lib/myposts.dart
@@ -320,7 +320,7 @@ class _MyEditPostState extends State<MyEditPost> {
           _phone.text,
           _email.text,
           _location.text,
-          int.parse(_price.text),
+          double.parse(_price.text),
           _currency,
           _listImage,
           _listRemoveImage);
